+++ map_reduce/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "map_reduce"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+rayon = "1.8.0"
+bincode = "1.3.3"
+tonic-build = "0.10.2"
+tonic = "0.10.2"
+prost = "0.12.1"
+tokio = { version = "1.34.0", features = ["full"] }
+
+[build-dependencies]
+tonic-build = {version="0.10.2", features = ["prost"]}
+
+
+
