+++ map_reduce/src/worker_service.rs
